--- kube-runtime/src/events/mod.rs
@@ -1,8 +1,0 @@
-//! Publishes events for objects
-#![allow(clippy::module_name_repetitions)]
-
-pub use event::{ControllerPodName, EventAction, EventNote, EventReason, EventSource, EventType, NewEvent};
-pub use recorder::EventRecorder;
-
-mod event;
-mod recorder;
